Time: 199 ms (11.79%), Space: 32.2 MB (94.53%) - LeetHub

+++ 1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts/1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts.cpp
@@ -14,6 +14,6 @@ public:
             max_c = max(max_c, verticalCuts[i]-verticalCuts[i-1]);
         }
         
-        return ((long)max_h * max_c) % (1000000007);
+        return (((long)max_h * (long)max_c) % (1000000007));
     }
 };
